+++ azure/rs_azure_compute/compute_test_cat.rb
@@ -1,0 +1,62 @@
+name 'Azure Compute Test CAT'
+rs_ca_ver 20161221
+short_description "Azure Compute - Test CAT"
+import "sys_log"
+import "plugins/rs_azure_compute"
+
+parameter "subscription_id" do
+  like $rs_azure_compute.subscription_id
+end
+
+permission "read_creds" do
+  actions   "rs_cm.show_sensitive","rs_cm.index_sensitive"
+  resources "rs_cm.credentials"
+end
+
+output "vm_ids" do
+  label "VM ids"
+end
+
+resource "my_availability_group", type: "rs_azure_compute.availability_set" do
+  name @@deployment.name
+  resource_group "rs-default-centralus"
+  location "Central US"
+  sku do {
+    "name" => "Aligned"
+  } end
+  properties do {
+      "platformUpdateDomainCount" => 5,
+      "platformFaultDomainCount" => 3
+  } end
+end
+
+operation "launch" do
+ description "Launch the application"
+ definition "launch_handler"
+ output_mappings do {
+    $vm_ids => $vms
+  } end
+end
+
+define launch_handler(@my_availability_group) return @my_availability_group,$vms do
+  call start_debugging()
+  provision(@my_availability_group)
+  $vms = to_s(@my_availability_group.virtualmachines)
+  call sys_log.detail("vms:" + to_s($vms))
+  call stop_debugging()
+end
+
+define start_debugging() do
+  if $$debugging == false || logic_and($$debugging != false, $$debugging != true)
+    initiate_debug_report()
+    $$debugging = true
+  end
+end
+
+define stop_debugging() do
+  if $$debugging == true
+    $debug_report = complete_debug_report()
+    call sys_log.detail($debug_report)
+    $$debugging = false
+  end
+end