+++ aws/rs_aws_rds/changelog.md
@@ -1,5 +1,10 @@
 RDS Plugin changelog
 
+v1.1
+-----
+- bugfix: change some field types that were "number" values to "string" values
+- bugfix: fix hrefs
+
 v1.0
 -----
 - initial release